+++ src/main/kotlin/org/yelog/ideavim/flash/ConfigUI.kt
@@ -6,10 +6,12 @@ import javax.swing.JTextField
 class ConfigUI {
     lateinit var rootPanel: JPanel
     private lateinit var charactersTF: JTextField
-    private lateinit var labelBgTF: JTextField
     private lateinit var labelFgTF: JTextField
+    private lateinit var labelBgTF: JTextField
+    private lateinit var matchFgTF: JTextField
     private lateinit var matchBgTF: JTextField
-    private lateinit var matchBgOpacityTF: JTextField
+    private lateinit var matchNearestFgTF: JTextField
+    private lateinit var matchNearestBgTF: JTextField
 
     var characters: String?
         get() = charactersTF.text
@@ -35,14 +37,14 @@ class ConfigUI {
         set(c) {
             labelFgTF.text = Integer.toHexString(c)
         }
-    var matchBgOpacity: Float
+    var matchFg: Int
         get() = try {
-            matchBgOpacityTF.text.toFloat()
+            Integer.parseUnsignedInt(matchFgTF.text, 16)
         } catch (e: NumberFormatException) {
-            0.3f
+            0
         }
         set(c) {
-            matchBgOpacityTF.text = c.toString()
+            matchFgTF.text = Integer.toHexString(c)
         }
     var matchBg: Int
         get() = try {
@@ -52,5 +54,23 @@ class ConfigUI {
         }
         set(c) {
             matchBgTF.text = Integer.toHexString(c)
+        }
+    var matchNearestFg: Int
+        get() = try {
+            Integer.parseUnsignedInt(matchNearestFgTF.text, 16)
+        } catch (e: NumberFormatException) {
+            0
+        }
+        set(c) {
+            matchNearestFgTF.text = Integer.toHexString(c)
+        }
+    var matchNearestBg: Int
+        get() = try {
+            Integer.parseUnsignedInt(matchNearestBgTF.text, 16)
+        } catch (e: NumberFormatException) {
+            0
+        }
+        set(c) {
+            matchNearestBgTF.text = Integer.toHexString(c)
         }
 }