+++ src/main/kotlin/org/yelog/ideavim/flash/action/Search.kt
@@ -16,38 +16,38 @@ class Search : Finder {
     // 记录可视区域的相对文档开头的位置
     private lateinit var visibleRange: TextRange
 
-    // 记录当前搜索的字符串
-    private var searchString = ""
-
     override fun start(e: Editor, visibleString: String, visibleRange: TextRange): List<MarksCanvas.Mark>? {
         this.visibleString = visibleString
         this.visibleRange = visibleRange
-        this.searchString = ""
         return null
     }
 
-    override fun input(e: Editor, c: Char, lastMarks: List<MarksCanvas.Mark>): List<MarksCanvas.Mark> {
+    override fun input(
+        e: Editor,
+        c: Char,
+        lastMarks: List<MarksCanvas.Mark>,
+        searchString: String
+    ): List<MarksCanvas.Mark> {
         if (lastMarks.any { it.keyTag.contains(c) }) {
             // hit a tag
             return advanceMarks(c, lastMarks);
         } else {
             // keep on searching
-            this.searchString += c
             val caretOffset = e.caretModel.offset
-            val offsets = visibleString.findAll(this.searchString, !this.searchString.contains(Regex("[A-Z]")))
+            val offsets = visibleString.findAll(searchString, !searchString.contains(Regex("[A-Z]")))
                 .map { it + visibleRange.startOffset }
                 .sortedBy { abs(it - caretOffset) }
                 .toList()
 
             val nextCharList =
-                offsets.map { this.visibleString[it - visibleRange.startOffset + this.searchString.length] }.distinct()
+                offsets.map { this.visibleString[it - visibleRange.startOffset + searchString.length] }.distinct()
             var remainCharacter = UserConfig.getDataBean().characters
             for (s in nextCharList) {
                 remainCharacter = remainCharacter.replace(s.toString(), "", true)
             }
             val tags = KeyTagsGenerator.createTagsTree(offsets.size, remainCharacter)
             return offsets.zip(tags)
-                .map { MarksCanvas.Mark(it.second, it.first, this.searchString.length) }
+                .map { MarksCanvas.Mark(it.second, it.first, searchString.length) }
                 .toList()
         }
     }